Jarvis/requirements.txt
discord.py
PyPDF2
qrcode
Pillow
fpdf
supabase
tabulate
google.generativeai
